refactor(address): inline customer address form (no modal) + fix State/Country crop
- Replace CustomerAddressModal (Dialog) with an inline CustomerAddressForm
  rendered in place inside the Addresses card on the customer detail page —
  matches the app's inline-editing convention; no overlay.
- Fix the State/Country cropping ("Calif…", "Unit…"): the address sub-fields
  were a 4-across grid (md:3) squeezed into a narrow modal (~140px each), so
  the Autocomplete ellipsized. Now a 2-up layout (City|State, Postal|Country)
  across all three address forms (customer, VendorForm, CompanyProfileCard),
  giving each field room.

No behavior change to validation/save; CustomerAddressModal deleted (no other
callers). tsc/lint/tests green (560 passing).
