+++ components/customers/index.ts
@@ -1,4 +1,4 @@
 export { default as CustomerStatusChip } from './CustomerStatusChip';
 export { default as CustomerForm } from './CustomerForm';
 export { default as CustomerContactModal } from './CustomerContactModal';
-export { default as CustomerAddressModal } from './CustomerAddressModal';
+export { default as CustomerAddressForm } from './CustomerAddressForm';
